+++ test/exif_test.cpp
@@ -18,7 +18,7 @@ TEST(Exif, Read)
     struct image* image = image_create();
     process_exif(image, data.data(), data.size());
 
-    EXPECT_EQ(image->num_info, 7);
+    EXPECT_EQ(image->num_info, static_cast<size_t>(7));
     EXPECT_STREQ(image->info[0].value, "2024:07:06 12:31:44");
     EXPECT_STREQ(image->info[1].value, "Google");
     EXPECT_STREQ(image->info[2].value, "Pixel 7");
@@ -35,10 +35,10 @@ TEST(Exif, Fail)
     struct image* image = image_create();
 
     process_exif(image, nullptr, 0);
-    EXPECT_EQ(image->num_info, 0);
+    EXPECT_EQ(image->num_info, static_cast<size_t>(0));
 
     process_exif(image, reinterpret_cast<const uint8_t*>("abcd"), 4);
-    EXPECT_EQ(image->num_info, 0);
+    EXPECT_EQ(image->num_info, static_cast<size_t>(0));
 
     image_free(image);
 }
